requirements.txt
streamlit
langchain
langchain-google-genai # NEW: For Gemini LLM and Embeddings
langchain-community
langchain-text-splitters
pydantic
python-dotenv
faiss-cpu
sentence-transformers # Keep this for general utility/sub-dependency reasons
pdfplumber
tiktoken
